Degree generator changed

+++ zbrozonoidEngine/Interfaces/IBall.cs
@@ -20,7 +20,7 @@ namespace zbrozonoidEngine.Interfaces
     {
         int OffsetX { get; set; }
         int OffsetY { get; set; }
-        double Degree { get; set; }
+        int Degree { get; set; }
         int DirectionX { get; set; }
         int DirectionY { get; set; }
         int Iteration { get; set; }
